refactor: replace barber and staff cards with a unified member card component, update member data retrieval to include avatarUrl
